requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.19.2
matplotlib>=3.3.2
Pillow>=8.0.0
scipy>=1.5.2
tqdm>=4.65.0
opencv-python>=4.8.0
scikit-image>=0.19.0